requirements.txt
Django==5.0.7
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
Pillow==10.0.0
python-decouple==3.8